Move cache clearing to notification
