+++ src/Umbraco.Community.CSPManager/Composing/CspManagerComposer.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Services;
 using Umbraco.Cms.Core.Notifications;
 using Umbraco.Community.CSPManager.Notifications;
+using Umbraco.Community.CSPManager.Notifications.Handlers;
 
 public sealed class CspManagerComposer : IComposer
 {
@@ -31,5 +32,6 @@ public sealed class CspManagerComposer : IComposer
 		});
 
 		builder.AddNotificationHandler<ServerVariablesParsingNotification, ServerVariablesHandler>();
+		builder.AddNotificationHandler<CspSavingNotification, CspSavingNotificationHandler>();
 	}
 }
