+++ src/Umbraco.Community.CSPManager/Services/CspService.cs
@@ -3,29 +3,26 @@
 using Cms.Core.Events;
 using Cms.Core.Hosting;
 using Cms.Infrastructure.Scoping;
-using Cms.Core.Web;
 using Models;
 using Umbraco.Extensions;
 using Umbraco.Cms.Core.Cache;
+using Umbraco.Community.CSPManager.Notifications;
 
 public class CspService : ICspService
 {
 	private readonly IEventAggregator _eventAggregator;
 	private readonly IHostingEnvironment _hostingEnvironment;
-	private readonly IUmbracoContextAccessor _umbracoContextAccessor;
 	private readonly IScopeProvider _scopeProvider;
 	private readonly IAppPolicyCache _runtimeCache;
 
 	public CspService(
 		IEventAggregator eventAggregator, 
 		IHostingEnvironment hostingEnvironment,
-		IUmbracoContextAccessor umbracoContextAccessor, 
 		IScopeProvider scopeProvider,
 		AppCaches appCaches)
 	{
 		_eventAggregator = eventAggregator;
 		_hostingEnvironment = hostingEnvironment;
-		_umbracoContextAccessor = umbracoContextAccessor;
 		_scopeProvider = scopeProvider;
 		_runtimeCache = appCaches.RuntimeCache;
 	}
@@ -82,9 +79,7 @@ public class CspService : ICspService
 
 		scope.Complete();
 
-		string cacheKey = definition.IsBackOffice ? CspConstants.BackOfficeCacheKey : CspConstants.FrontEndCacheKey;
-
-		_runtimeCache.ClearByKey(cacheKey);
+		_eventAggregator.Publish(new CspSavingNotification(definition));
 
 		return definition;
 	}
